Fix ScoopTrigger baseline race, WaveTrigger time window, PickupDeviceTrigger buffer clearing
ScoopTrigger: Add baselineReadings counter to skip impulse detection
during the first 10 readings, letting the EMA baseline converge before
evaluating impulses. Prevents false triggers from stale GRAVITY_EARTH.

WaveTrigger: Change timeWindowMillis from Float to Long. Only update
lastProximityEventTime on the Near transition, so the time window correctly
measures the full gesture duration (Near→Far) instead of resetting on
every continuous-sensor reading.

PickupDeviceTrigger: Remove buffer clearing on PickedUp — the rolling
window self-heals naturally. Eliminates the 3-reading blind spot.

+++ sensey/src/main/kotlin/com/github/nisrulz/sensey/SenseyPluginRegistry.kt
@@ -113,7 +113,7 @@ class SenseyPluginRegistry {
     }
 
     fun wavePlugin(
-        timeWindowMillis: Float = 1000f,
+        timeWindowMillis: Long = 1000L,
         debounceMillis: Long = 1000L,
         dispatcher: (WaveEvent) -> Unit,
     ) {
