+++ sensey/src/main/kotlin/com/github/nisrulz/sensey/gesture/GesturePlugins.kt
@@ -168,7 +168,7 @@ fun wristTwistPlugin(
     )
 
 fun wavePlugin(
-    timeWindowMillis: Float = 1000f,
+    timeWindowMillis: Long = 1000L,
     debounceMillis: Long = 1000L,
     dispatcher: (WaveEvent) -> Unit,
 ): GesturePlugin =
